moved shared.h into main folder
it was only shared for mac, becaused it was used by 2 tools
