+++ tools/linuxdeployqt/shared.cpp
