+++ tools/linuxdeployqt/shared.h
